fetch only one product

+++ src/main/java/com/example/fines_saveurs/dao/ProductDao.java
@@ -2,6 +2,6 @@ package com.example.fines_saveurs.dao;
 
 import com.example.fines_saveurs.model.Product;
 
-public interface ProductDao extends GenericDao<Product, Long> {
+public interface ProductDao extends GenericDao<Product, Integer> {
 
 }
